+++ xenian/bot/commands/__init__.py
@@ -7,7 +7,7 @@ from .urban_dictionary import *
 from .roll import *
 from .download import *
 from .reverse_image_search import *
-from .danbooru import *
+from .animedatabases import *
 from .decide import *
 from .calculator import *
 from .group_management import *