+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 - Replace Yandex- with Google-Translate
 - Remove ``/danbooru_latest`` and replace ``/danbooru_search`` with ``/danbooru``
 - Add more danbooru like services. Instead of just danbooru we get safebooru, yandere and konachan too
+- Rename Danbooru class to AnimeDtabases because it fits better§
 
 
 2.4.0 (2019-01-05)